feat: "no-mutable-public-field" rule to "no-mutable-public-property-of-construct"
