+++ src/index.ts
@@ -7,7 +7,7 @@ import { noConstructInInterface } from "./rules/no-construct-in-interface";
 import { noConstructStackSuffix } from "./rules/no-construct-stack-suffix";
 import { noImportPrivate } from "./rules/no-import-private";
 import { noMutablePropsInterface } from "./rules/no-mutable-props-interface";
-import { noMutablePublicFields } from "./rules/no-mutable-public-fields";
+import { noMutablePublicPropertyOfConstruct } from "./rules/no-mutable-public-property-of-construct";
 import { noParentNameConstructIdMatch } from "./rules/no-parent-name-construct-id-match";
 import { noPublicClassFields } from "./rules/no-public-class-fields";
 import { noVariableConstructId } from "./rules/no-variable-construct-id";
@@ -25,7 +25,7 @@ const rules = {
   "pascal-case-construct-id": pascalCaseConstructId,
   "require-passing-this": requirePassingThis,
   "no-variable-construct-id": noVariableConstructId,
-  "no-mutable-public-fields": noMutablePublicFields,
+  "no-mutable-public-property-of-construct": noMutablePublicPropertyOfConstruct,
   "no-mutable-props-interface": noMutablePropsInterface,
   "construct-constructor-property": constructConstructorProperty,
   "require-jsdoc": requireJSDoc,
@@ -65,7 +65,7 @@ const recommended = createFlatConfig({
   "cdk/pascal-case-construct-id": "error",
   "cdk/require-passing-this": ["error", { allowNonThisAndDisallowScope: true }],
   "cdk/no-variable-construct-id": "error",
-  "cdk/no-mutable-public-fields": "warn",
+  "cdk/no-mutable-public-property-of-construct": "warn",
   "cdk/no-mutable-props-interface": "warn",
   "cdk/construct-constructor-property": "error",
 });
@@ -81,7 +81,7 @@ const strict = createFlatConfig({
   "cdk/pascal-case-construct-id": "error",
   "cdk/require-passing-this": "error",
   "cdk/no-variable-construct-id": "error",
-  "cdk/no-mutable-public-fields": "error",
+  "cdk/no-mutable-public-property-of-construct": "error",
   "cdk/no-mutable-props-interface": "error",
   "cdk/construct-constructor-property": "error",
   "cdk/require-jsdoc": "error",
